Some Bugs and Fixes
The registration app now only takes 6 images, but only keeps three. If
the server is unavailable at sending of data, then the error message is
more clear. Also fixed a bug where if there was not enough images
taken, then the application would become unresponsive. We still have
the issue of live streaming of facial data is not shown upon second
traversal, but we are looking into fixing this.